pyproject.toml
[project]
name = "arxiv-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "fastapi>=0.116.1",
    "google-genai>=1.29.0",
    "pydantic-settings>=2.10.1",
    "slowapi>=0.1.9",
    "uvicorn>=0.35.0",
]
